shiguang/requirement.txt
aliyun-python-sdk-core==2.13.5
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-kms==2.7.1
certifi==2019.6.16
chardet==3.0.4
crcmod==1.7
Django==2.2.4
django-mdeditor==0.1.16
django-redis==4.10.0
django-simpleui==2.6
idna==2.8
jmespath==0.9.4
mysqlclient==1.4.4
oss2==2.8.0
pycryptodome==3.8.2
pytz==2019.2
redis==3.3.7
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
